Message cleanup and porting issues
git-svn-id: file:///home/jj/hercules.svn/trunk@6985 956126f8-22a0-4046-8f4a-272fa8102e63

+++ cpu.c
@@ -1435,7 +1435,7 @@ int i;
     memset(&regs->blknam,SPACE,sizeof(regs->blknam));
     memset(&regs->blkver,SPACE,sizeof(regs->blkver));
     memset(&regs->blkend,SPACE,sizeof(regs->blkend));
-    regs->blkloc = swap_byte_U64((U64)regs);
+    regs->blkloc = swap_byte_U64((U64)((u_int)regs));
     regs->blksiz = swap_byte_U32((U32)sizeof(REGS));
     {
         char cputyp[32];
@@ -1795,7 +1795,7 @@ int     aswitch;
     {
         memcpy (&regs, oldregs, sizeof(REGS));
         free (oldregs);
-        regs.blkloc = swap_byte_U64((U64)&regs);
+        regs.blkloc = swap_byte_U64((U64)((u_int)&regs));
         regs.hostregs = &regs;
         if (regs.guestregs)
             regs.guestregs->hostregs = &regs;
